fix: 回退版本降级反模式 + 修复重采样dtype回退 + 新增反模式测试
回退反模式：
- v4.0a/v4.0ap: 删除 repair_single_track/repair_audio 的 try/except 降级包装
- 版本失败就报错，不能偷偷换成 v3.2a 的结果还说是 v4.0 的
- 删除 _v32a_repair_single_track / _v32a_repair_audio 导入

修复 DSP bug：
- v2.2/v2.3/v2.3a/v2.4a: 重采样 np.zeros 添加 dtype=y.dtype
- 修复 float32 内存优化失效问题（重采样后退回 float64 导致内存翻倍）

反模式测试：
- 新增 test_antipattern_version_fallback.py（22个测试）
- 严禁源代码中出现版本降级 fallback 关键词
- 严禁导入其他版本 repair 入口作为降级目标
- 严禁 algorithm_version 动态拼接 fallback 标签
- VERSION_TAG 必须是静态常量

276个回归测试 + 22个反模式测试全部通过

+++ backend/services/repair/repair_v2_3/core.py
@@ -304,7 +304,7 @@ def repair_audio(input_path: str, output_path: str, params: dict, progress_callb
         if progress_callback:
             progress_callback(0.02, f"v2.3 重采样到 {working_sr//1000}kHz...")
         target_len = int(y.shape[1] * working_sr / sr)
-        y_new = np.zeros((y.shape[0], target_len))
+        y_new = np.zeros((y.shape[0], target_len), dtype=y.dtype)
         for ch in range(y.shape[0]):
             resampled = resample_poly(y[ch], working_sr, sr)
             y_new[ch, :len(resampled)] = resampled[:target_len]
